% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agv.R
\name{agv}
\alias{agv}
\title{agv}
\usage{
agv(x, y, k, rep, H, method, dist)
}
\arguments{
\item{x}{is n*p matrix}

\item{y}{is n-dim vector}

\item{k}{is dimension of S (augmentation predictor), default p/10}

\item{rep}{is number of repitations of augmentation, default 10}

\item{H}{is number of slices (for inverse regression)}

\item{method}{is the working matrix function}

\item{dist}{is the distribution of S, normal(default) or runis}
}
\description{
main function
}
