% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ladle_split.R
\name{ladle_split}
\alias{ladle_split}
\title{ladle_split}
\usage{
ladle_split(
  x = x,
  y = y,
  nslices = nslices,
  nsplit = nsplit,
  method = method,
  order = order,
  criterion = criterion,
  n0 = n0
)
}
\arguments{
\item{nslices, }{parameter needed in the SIR methods}

\item{nsplit, }{nsplit means we want to split the whole samples into nsplit parts}

\item{method, }{AFM, pca, cca, ica, sir, save, dr, DEE, kernel sir.}

\item{order, }{parameter needed in the SIR methods}

\item{criterion, }{criterion = 1 means the first B_k combination methods, means the total B_k dots each split part of B_ik, and sum
criterion = 2 means each split parts dots (even split is needed)
criterion = 3 means criterion1 + criterion2 and then average them}

\item{n0, }{n0 means to repeat split method n0 times}
}
\description{
The main function in splitLadle method
}
