% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ladle_weight.R
\name{ladle_weight}
\alias{ladle_weight}
\title{ladle_weight}
\usage{
ladle_weight(
  x = x,
  y = y,
  nslices = nslices,
  s = s,
  method = method,
  order = order,
  wt = wt
)
}
\arguments{
\item{x, }{predictors}

\item{y, }{response}

\item{nslices, }{parameter needed in the SIR methods}

\item{method, }{AFM, pca, cca, ica, sir, save, dr, DEE, kernel sir.}

\item{order, }{used in SIR}

\item{wt, }{the given weight type 'normal','bernoulli'}

\item{nsplit, }{nsplit means we want to split the whole samples into nsplit parts}
}
\description{
ladle weight methods
}
