% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ladle_prepare.R
\name{mhat_dr}
\alias{mhat_dr}
\title{mhat_dr}
\usage{
mhat_dr(x, y, H)
}
\arguments{
\item{x, }{a matrix}

\item{y, }{a vector or matrix}

\item{H, }{a slice}
}
\description{
mhat for directional regression
}
