% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ladle_prepare.R
\name{mhat_ksir}
\alias{mhat_ksir}
\title{mhat_ksir}
\usage{
mhat_ksir(x = x, phi = phi, y = y, nslices = nslices)
}
\arguments{
\item{x, }{the original predictor}

\item{phi, }{the transformed predictor -in kernel space}

\item{y, }{the original response}

\item{nslices, }{slices number}

\item{Sigma, }{a matrix}
}
\description{
mhat for kernel SIR
}
