% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ladle_prepare.R
\name{mppower}
\alias{mppower}
\title{mppower}
\usage{
mppower(matrix, power, ignore)
}
\arguments{
\item{matrix, }{a matrix}

\item{power, }{a number}

\item{ignore, }{criterion=ignore}
}
\description{
Moore-Penrose type power
}
