% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRE_RE_BIC.R
\name{RRE_RE_BIC}
\alias{RRE_RE_BIC}
\title{RRE_RE_BIC}
\usage{
RRE_RE_BIC(x = x, y = y, method = method, nslices = nslices)
}
\arguments{
\item{x, }{predictors}

\item{y, }{response}

\item{method, }{mtthod choices}

\item{nslices, }{used in SIR}
}
\description{
Xia et al(2015), Luo et al(2009), Zhu et al(2006,2010)
}
