% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agv.R
\name{runis}
\alias{runis}
\title{runis}
\usage{
runis(n, p)
}
\arguments{
\item{n, }{items}

\item{p, }{dimension}
}
\description{
generate RV from a uniform distribution
on a hyper-sphere centered at the origin and with radius sqrt p
}
