% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDRR.R
\name{TDRR}
\alias{TDRR}
\title{TDRR}
\usage{
TDRR(x = x, y = y, method = method, nslices = nslices, tao)
}
\arguments{
\item{x, }{predictors}

\item{y, }{response}

\item{method, }{mtthod choices}

\item{nslices, }{used in SIR}

\item{tao, }{tunning parameters}
}
\description{
zhu et.al(2020)
}
