% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ladle.R
\name{yeweiss}
\alias{yeweiss}
\title{yeweiss}
\usage{
yeweiss(
  x = x,
  y = y,
  nslices = nslices,
  nboot = nboot,
  method = method,
  order = order,
  delta = delta
)
}
\arguments{
\item{x, }{predictors}

\item{y, }{response}

\item{nslices, }{parameter needed in the SIR methods}

\item{nboot, }{bootstrap 次数}

\item{method, }{AFM, pca, cca, ica, sir, save, dr, DEE, kernel sir.}

\item{order, }{parameter needed in the SIR methods}

\item{delta}{is the scalar to select the threshold for "small" bootstrap variability}
}
\description{
The original bootstrap ladle consider eigenvectors only
}
